+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added `From` implementation for `asset::Symbol` type
+- Added support for almost-default construction of various request types
 - Converted `Account::daytrade_count` to `u64`
 - Decreased tracing verbosity by one level
 - Bumped `num-decimal` dependency to `0.2`